pyproject.toml
[tool.pdm]

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Ciastelix", email = "mateusz.zwierzynski@hotmail.com"},
]
dependencies = [
    "fastapi>=0.95.1",
    "sqlalchemy>=1.4.47",
    "python-dotenv>=1.0.0",
    "dependency-injector>=4.41.0",
    "uvicorn>=0.21.1",
    "pydantic-sqlalchemy>=0.0.9",
    "pytest>=7.3.1",
]
requires-python = ">=3.11"
license = {text = "MIT"}
